ml_service/requirements_chronos.txt
fastapi==0.115.0
uvicorn==0.30.0
chronos-forecasting==1.3.0
torch>=2.6.0
numpy>=1.26.4
pydantic>=2.7.0